Refactory Key interface into wildfly-clustering-ee-spi.
