+++ clustering/ee/infinispan/src/main/java/org/wildfly/clustering/ee/infinispan/scheduler/ScheduleLocalEntriesTask.java
@@ -30,7 +30,7 @@ import java.util.stream.Stream;
 
 import org.infinispan.Cache;
 import org.infinispan.context.Flag;
-import org.wildfly.clustering.ee.cache.Key;
+import org.wildfly.clustering.ee.Key;
 import org.wildfly.clustering.infinispan.spi.distribution.Locality;
 
 /**
