+++ clustering/ee/spi/src/main/java/org/wildfly/clustering/ee/Key.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.ee.cache;
+package org.wildfly.clustering.ee;
 
 /**
  * A cache key for a given identifier
