+++ clustering/ejb/infinispan/src/main/java/org/wildfly/clustering/ejb/infinispan/BeanGroupKey.java
@@ -22,7 +22,7 @@
 
 package org.wildfly.clustering.ejb.infinispan;
 
-import org.wildfly.clustering.ee.cache.Key;
+import org.wildfly.clustering.ee.Key;
 
 /**
  * @author Paul Ferraro
